requirements.txt
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
colorama==0.4.6
python-dotenv==1.0.0
Flask==2.3.3
pyinstaller==6.3.0
Pillow==10.1.0
